Se adeacuan los ccontroller e interfaces de las API´s

+++ src/main/java/com/edu/mx/rest/est/siest12/controller/BloodController.java
@@ -41,7 +41,12 @@ public class BloodController {
 
     @DeleteMapping("/delete/{idBlood}")
     public String deleteBlood(@PathVariable int idBlood){
-        bloodService.deleteBlood(idBlood);
+        Blood objBlood = bloodService.getBlood(idBlood);
+
+        if (objBlood != null){
+            bloodService.deleteBlood(idBlood);
+        }
+
         return "El registro se a liminado de forma correcta en tabla SANGRE";
     }
